backend/requirements.txt
# FastAPI and server
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18

# Database
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==44.0.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0

pyjwt==2.10.1

# HTTP client for external APIs
httpx==0.28.1
requests==2.32.3

# Environment variables
python-dotenv==1.0.1

# Data validation
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# Utilities
python-dateutil==2.9.0

# Logging
python-json-logger==3.2.1

# Caching (optional but recommended)
redis==5.2.1
